moved styled components to a separate file
